[#36]feature-#36:페이징 처리 피드백 반영
